Restructure package format, add Composite Packages (#21)
* Restructure package format, add Composite Packages

This PR re-structures the format of the "package" structure,
which also has implications for package-manifest.toml files.

What did we change, and why?

- Unified internal/external packages: There is information about
packages which is shared between internal and external packages,
so they've been pulled into the same structure. The "input" for a
package is now described by "PackageSource", which can be one
of "local" (you're building the package), "prebuilt" (you're downloading
the package), "composite" (you're combining packages") or "manual"
(you're supplying the package yourself).

- The boolean "zone" flag has been updated to an enum, called
"PackageOutput". This enum describes the different outputs that
one should expect from a package.

- The "Target" evaluation information has been moved to this package
from the primary Omicron repository. This makes it easier to act on the
'only_for_target' and 'intermediate_only' fields in a useful way.

* Review feedback

* Async append_dir_all