Add support for downloading prebuilt blobs from Buildomat
Allow packaging TOMLs to define one or more `buildomat_blobs` that refer to a
Buildomat artifact. This artifact need not be a zone image.

The existing S3 blob type is not changed, and buildomat blobs are optional, so
this should be a non-breaking change for existing package manifests.

Tested: cargo test; picked up the changes into Propolis's packaging utility
and verified that they can be used to pick up guest firmware images produced
by Buildomat.