ensure blob directory exists in archives (#77)
The helios-omicron-brand code that unpacks zone images requires all file
entries in an archive to be preceded by entries that create those files'
directories. Ensure the packaging code adds the requisite entries for
downloaded blobs (as it already does for Rust programs).

Propolis will need a patch release containing this fix to produce
functioning Omicron zone images, so I've bumped the package version to
0.12.2 in this commit; let me know if I should do that in a separate PR.

Tested by building a Propolis package using this commit and verifying
that (a) tar -tzf prints the appropriate output structure, and (b)
hotpatching an Omicron dev cluster with this zone image allows instances
to start (images created without the fix produce a file-not-found error
when sled-agent tries to start them).